requirements.txt
Flask==2.0.1
googletrans==3.1.0a0
google-cloud-texttospeech==2.12.0
gTTS==2.3.1
